RagChat/requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#streamlit
streamlit

#langchain community
langchain-community
pypdf

#vector database
faiss-cpu
rank_bm25
pymupdf
sentence-transformers